completed the ide, at least the linux build of it
copied in the latest libmaple via copy-to-ide, modified boards.txt to reflect the vendor product id change to 1eaf:0003, tested that everything builds and runs from build/linux, moved dfu-util to dist/tools where it belongs, thats also where the ide looks for it. the current version doesnt seem to throw nasty errors anymore after using the serial to reset and dfu to load to new code, which means the timing is right.