Removing libmaple dependencies from repo.
The IDE repo was duplicating a bunch of libmaple files.  These are
removed; it's now the the responsibility of whoever compiles the IDE
to put libmaple dependencies in the correct places.

The libmaple file support/scripts/copy-to-ide is intended to make
doing this easy.  The README has been updated to reflect its usage.